requirements.txt
streamlit==1.39.0
gspread==6.1.2
google-auth==2.35.0
gspread-dataframe==3.3.1
pandas==2.2.2
python-dateutil==2.9.0.post0
